requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.6
pydantic==2.4.2
python-dotenv==1.0.1
openai==1.12.0
httpx==0.24.1
requests==2.31.0 